# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Estecapelli API — Preevaluación capilar
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SimpleOk {
  ok: boolean;
}

export interface AuthStatus {
  authenticated: boolean;
  demoPassword: boolean;
}

export interface AdminCredentials {
  password: string;
}

export interface LoginResult {
  ok: boolean;
  demoPassword: boolean;
}

export interface PhotoEntry {
  key: string;
  label: string;
  dataUrl: string;
  quality: string;
  createdAt: string;
}

export type PatientPhotoStatusStatus = typeof PatientPhotoStatusStatus[keyof typeof PatientPhotoStatusStatus];


export const PatientPhotoStatusStatus = {
  draft: 'draft',
  confirmed: 'confirmed',
  superseded: 'superseded',
  discarded: 'discarded',
} as const;

export type PatientPhotoStatusSource = typeof PatientPhotoStatusSource[keyof typeof PatientPhotoStatusSource];


export const PatientPhotoStatusSource = {
  camera: 'camera',
  upload: 'upload',
} as const;

export interface PatientPhotoStatus {
  id: string;
  key: string;
  label: string;
  status: PatientPhotoStatusStatus;
  source: PatientPhotoStatusSource;
  mimeType: string;
  sizeBytes: number;
  sha256: string;
  createdAt: string;
  /** @nullable */
  confirmedAt?: string | null;
  hasOriginal: boolean;
  hasAdjusted: boolean;
  editParams?: unknown;
}

export interface PatientPhotoStatusResult {
  ok: boolean;
  photos: PatientPhotoStatus[];
}

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  nuevo: 'nuevo',
  incompleto: 'incompleto',
  listo: 'listo',
  contactar: 'contactar',
  agendado: 'agendado',
  cerrado: 'cerrado',
} as const;

export interface LeadSummary {
  id: string;
  token: string;
  createdAt: string;
  updatedAt: string;
  name: string;
  phone: string;
  /** @nullable */
  documentId?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  city?: string | null;
  status: LeadStatus;
  consent?: boolean;
  photoCount: number;
  photoKeys?: string[];
  /** @nullable */
  hairLossTime?: string | null;
  /** @nullable */
  pattern?: string | null;
  /** @nullable */
  previousTreatment?: string | null;
  /** @nullable */
  norwood?: string | null;
  /** @nullable */
  appointmentAt?: string | null;
  /** @nullable */
  isDemo?: boolean | null;
}

export interface Lead {
  id: string;
  token: string;
  createdAt: string;
  updatedAt: string;
  name: string;
  phone: string;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  city?: string | null;
  status: LeadStatus;
  consent?: boolean;
  photoCount: number;
  photos: PatientPhotoStatus[];
  /** @nullable */
  hairLossTime?: string | null;
  /** @nullable */
  pattern?: string | null;
  /** @nullable */
  previousTreatment?: string | null;
  /** @nullable */
  symptoms?: string | null;
  /** @nullable */
  surgeryHistory?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  norwood?: string | null;
  /** @nullable */
  appointmentAt?: string | null;
  /** @nullable */
  isDemo?: boolean | null;
}

export type LeadStatsCounts = {[key: string]: number};

export interface LeadStats {
  total: number;
  counts: LeadStatsCounts;
}

export interface LeadPatch {
  status?: LeadStatus;
  notes?: string;
  norwood?: string;
  appointmentAt?: string;
}

export interface InvitationInput {
  name?: string;
  phone?: string;
}

export interface InvitationResult {
  ok: boolean;
  lead: LeadSummary;
  link: string;
}

export interface PatientInput {
  name: string;
  phone: string;
  documentId?: string;
  email?: string;
  age?: string;
  city?: string;
  hairLossTime?: string;
  pattern?: string;
  previousTreatment?: string;
  symptoms?: string;
  surgeryHistory?: string;
  consent: boolean;
  marketingConsent?: boolean;
}

export interface PatientResult {
  ok: boolean;
  lead: LeadSummary;
}

export type GetLeadsParams = {
status?: string;
search?: string;
};

export type GetLeads200 = {
  leads: LeadSummary[];
};