# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Estecapelli API — Preevaluación capilar
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of lead summaries
          content:
            application/json:
              schema:
                type: object
                required: [leads]
                properties:
                  leads:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeadSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/stats:
    get:
      operationId: getLeadStats
      tags: [leads]
      summary: Lead statistics (admin)
      responses:
        "200":
          description: Lead counts by status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}:
    get:
      operationId: getLeadById
      tags: [leads]
      summary: Get lead detail (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead detail with photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead permanently (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
        "403":
          description: Cannot delete (wrong status)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: patchLead
      tags: [leads]
      summary: Update lead status, notes, norwood, appointment (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadPatch"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}/photos/{photoId}:
    get:
      operationId: getLeadPhotoFile
      tags: [leads]
      summary: Stream a clinical photo to an authorized staff session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Private image stream
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SimpleOk:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string
    LoginResult:
      type: object
      required: [ok, demoPassword]
      properties:
        ok:
          type: boolean
        demoPassword:
          type: boolean
...
    PatientPhotoStatus:
      type: object
      required: [id, key, label, status, source, mimeType, sizeBytes, sha256, createdAt, hasOriginal, hasAdjusted]
      properties:
        id:
          type: string
        key:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [draft, confirmed, superseded, discarded]
        source:
          type: string
          enum: [camera, upload]
        mimeType:
          type: string
        sizeBytes:
          type: number
        sha256:
          type: string
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: ["string", "null"]
          format: date-time
        hasOriginal:
          type: boolean
        hasAdjusted:
          type: boolean
        editParams: {}
...
    LeadStatus:
      type: string
      enum: [nuevo, incompleto, listo, contactar, agendado, cerrado]
    LeadSummary:
      type: object
      required: [id, token, createdAt, updatedAt, name, phone, status, photoCount]
      properties:
        id:
          type: string
        token:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        name:
          type: string
        phone:
          type: string
        documentId:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/LeadStatus"
        consent:
          type: boolean
        photoCount:
          type: number
        photoKeys:
          type: array
          items:
            type: string
        hairLossTime:
          type: ["string", "null"]
        pattern:
          type: ["string", "null"]
        previousTreatment:
          type: ["string", "null"]
        norwood:
          type: ["string", "null"]
        appointmentAt:
          type: ["string", "null"]
        isDemo:
          type: ["boolean", "null"]
    Lead:
      type: object
      required: [id, token, createdAt, updatedAt, name, phone, status, photoCount, photos]
      properties:
        id:
          type: string
        token:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        name:
          type: string
        phone:
          type: string
        age:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/LeadStatus"
        consent:
          type: boolean
        photoCount:
          type: number
        photos:
          type: array
          items:
            $ref: "#/components/schemas/PatientPhotoStatus"
        hairLossTime:
          type: ["string", "null"]
        pattern:
          type: ["string", "null"]
        previousTreatment:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        surgeryHistory:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        norwood:
          type: ["string", "null"]
        appointmentAt:
          type: ["string", "null"]
        isDemo:
          type: ["boolean", "null"]
    LeadStats:
      type: object
      required: [total, counts]
      properties:
        total:
          type: number
        counts:
          type: object
          additionalProperties:
            type: number
    LeadPatch:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/LeadStatus"
        notes:
          type: string
        norwood:
          type: string
        appointmentAt:
          type: string
...