# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Estecapelli API — Preevaluación capilar
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from './leadStatus';
import type { PatientPhotoStatus } from './patientPhotoStatus';

export interface Lead {
  id: string;
  token: string;
  createdAt: string;
  updatedAt: string;
  name: string;
  phone: string;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  city?: string | null;
  status: LeadStatus;
  consent?: boolean;
  photoCount: number;
  photos: PatientPhotoStatus[];
  /** @nullable */
  hairLossTime?: string | null;
  /** @nullable */
  pattern?: string | null;
  /** @nullable */
  previousTreatment?: string | null;
  /** @nullable */
  symptoms?: string | null;
  /** @nullable */
  surgeryHistory?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  norwood?: string | null;
  /** @nullable */
  appointmentAt?: string | null;
  /** @nullable */
  isDemo?: boolean | null;
}