# lib/api-zod/src/generated/types/patientPhotoStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Estecapelli API — Preevaluación capilar
 * OpenAPI spec version: 0.1.0
 */
import type { PatientPhotoStatusSource } from './patientPhotoStatusSource';
import type { PatientPhotoStatusStatus } from './patientPhotoStatusStatus';

export interface PatientPhotoStatus {
  id: string;
  key: string;
  label: string;
  status: PatientPhotoStatusStatus;
  source: PatientPhotoStatusSource;
  mimeType: string;
  sizeBytes: number;
  sha256: string;
  createdAt: Date;
  /** @nullable */
  confirmedAt?: Date | null;
  hasOriginal: boolean;
  hasAdjusted: boolean;
  editParams?: unknown;
}